chore(file structure): reorganised the file structure of arena- or term-related functions
