+++ kernel/src/term/mod.rs
@@ -1,0 +1,3 @@
+pub mod arena;
+pub mod builders;
+pub mod calculus;
